refactor(app): move the e2e fixtures out of the server action
app/_lib/actions.ts was 569 lines, 173 of them browser-test fixtures --
thirty percent of the "use server" module that queries GitHub with the
token, including one fixture that throws on purpose and two that keep
state between calls. Test scaffolding does not belong in the most
security-sensitive file here, and its presence made the real search path
harder to read than it needed to be.

The fixtures now live in app/_lib/e2eCommitSearchFixtures.ts. actions.ts
is 379 lines and calls into it twice.

Two things had to move first. commitSearchError and the two empty-search
messages are shared between the production path and the fixtures, and a
"use server" module cannot export a synchronous helper, so they moved to
commitTypes.ts beside toCommitSearchSuccess -- both constructors of the
union now live with the union. Sharing rather than copying them matters
for the same reason commitSearchMocksEnabled has one definition: two
copies would let the real path and the fixture path describe the same
outcome in different words.

The reject-once hook could not move as a fixture result, because its
whole purpose is to throw rather than return. It became
shouldRejectCommitSearchOnce, which owns the decision and the state while
actions.ts still owns the throw.

The browser suite is the regression test, since no fixture has a unit
test to lean on. All 51 specs pass, including the four that depend on
state surviving between two requests in one process: the two partial-retry
journeys, the failed-retry journey, and the Server Action rejection.

Coverage of actions.ts rose from 76.54% to 93.33% statements and 71.08%
to 86.25% branches, because it is no longer measured against lines no
unit test could reach. The aggregate moved with it, so the floors were
raised to 96/93/98/97 -- leaving them at 93/89/98/94 would have left
three and a half points of slack, which is room for real regression to
pass. The new module is excluded on the same footing as the other files
the gate covers by running rather than importing.

Re-proved at the new floors: a two-function module no test imports still
fails the gate, on branches at 92.77% against 93.

Closes #185

Entire-Checkpoint: 01M08AN7MZ8TJPPF64H5BA8SXZ
