+++ CHANGELOG.md
@@ -41,6 +41,8 @@ This project follows a lightweight, human-curated changelog. Keep the newest cha
 
 ### Changed
 
+- The browser-test fixtures moved out of the server action that queries GitHub. Thirty percent of that file — 173 of 569 lines — was test doubles for the reserved `e2e-*` usernames, including one that throws on purpose and two that keep state between calls, all sitting in the module that holds the token and the real search path. They now live in their own module, which nothing but the browser suite reaches. Nothing about the app changed; the file that matters is a third shorter and reads as the production path it is, and coverage of it rose from 76.54% to 93.33% because it is no longer measured against lines no unit test could ever run. The coverage floors were raised to match the new measurement rather than left with three and a half points of slack.
+
 - Coverage now measures the directories the unit suite collects from rather than only the files it happens to load, so a module no test imports counts as zero instead of vanishing from the report. The thresholds were a floor that could catch coverage slipping in code already under test while missing code that arrived with no test at all — the more common way coverage decays. Seven files are excluded, each one the gate covers by running it rather than importing it: the three metadata image routes the production build prerenders, and the four command-line entry points the gate and the production workflows execute, every one a thin shell over a module that is unit tested. The thresholds did not move, which is the evidence the exclusions are right rather than convenient: scoped and excluded, the numbers are the same to two decimal places as before.
 
 - Node type definitions now match the Node the project runs on, and the Vitest config is loaded as the ES module it is. `@types/node` was pinned to major 22 while `.nvmrc` and the package engines both say 24, so the types described a different runtime than the one executing — the kind of mismatch that reports an API as missing when it exists, or the reverse. The config file moved to `vitest.config.mts`, which is what its `import.meta.url` needs: as `.ts` it was loaded through a CommonJS compatibility path that printed a deprecation warning on every test run, and that path is slated to become an error rather than a warning. The check that keeps unit collection scoped reads the file by name and moved with it in the same change, so the rename could not leave the gate looking for a file that no longer exists.
