+++ app/_lib/actions.ts
@@ -1,8 +1,18 @@
 "use server";
 
 import { Octokit } from "octokit";
-import { toCommitSearchSuccess } from "./commitTypes";
-import type { CommitData, CommitErrorKind, CommitInfo, CommitSearchFailure } from "./commitTypes";
+import {
+  EMPTY_COMMIT_SEARCH_MESSAGE,
+  INCOMPLETE_COMMIT_SEARCH_MESSAGE,
+  commitSearchError,
+  toCommitSearchSuccess,
+} from "./commitTypes";
+import type { CommitData, CommitErrorKind, CommitInfo } from "./commitTypes";
+import {
+  E2E_SLOW_SEARCH_DELAY_MS,
+  getE2eCommitSearchResult,
+  shouldRejectCommitSearchOnce,
+} from "./e2eCommitSearchFixtures";
 import { getCachedCommitSearch, setCachedCommitSearch } from "./commitSearchCache";
 import { coalesceCommitSearch } from "./commitSearchInFlight";
 import {
@@ -10,7 +20,6 @@ import {
   COMMIT_SEARCH_RATE_LIMIT_WINDOW_MS,
   allowCommitSearch,
 } from "./commitSearchRateLimit";
-import { commitSearchMocksEnabled } from "./e2eCommitSearchMocks";
 import { githubProfileUrl } from "./githubUrls";
 import { logger } from "./logger";
 import { getSearchClientKey } from "./searchClientKey";
@@ -21,25 +30,6 @@ const octokit = new Octokit({
 });
 
 const GITHUB_SEARCH_TIMEOUT_MS = 10_000;
-const E2E_COMMIT_SEARCH_MOCKS_ENABLED = commitSearchMocksEnabled();
-const E2E_SLOW_SEARCH_DELAY_MS = 750;
-const E2E_REJECT_ONCE_USERNAME_PREFIX = "e2e-reject-once-";
-const e2eRejectedUsernames = new Set<string>();
-const E2E_INCOMPLETE_ONCE_USERNAME_PREFIX = "e2e-incomplete-once-";
-const E2E_INCOMPLETE_THEN_ERROR_USERNAME_PREFIX = "e2e-incomplete-then-error-";
-const e2eServedPartialUsernames = new Set<string>();
-const EMPTY_COMMIT_SEARCH_MESSAGE =
-  "No public commits found for this user (or indexing is delayed).";
-const INCOMPLETE_COMMIT_SEARCH_MESSAGE =
-  "GitHub could not finish this search, so no commits were returned yet.";
-
-function commitSearchError(
-  error: string,
-  errorKind: CommitErrorKind,
-  incomplete = false,
-): CommitSearchFailure {
-  return { found: false, error, errorKind, ...(incomplete ? { incomplete } : {}), commits: [] };
-}
 
 type GitHubErrorDetails = {
   message?: string;
@@ -57,180 +47,6 @@ function parseNonNegativeInteger(value: unknown): number | undefined {
 
   const parsedValue = Number(value);
   return Number.isSafeInteger(parsedValue) ? parsedValue : undefined;
-}
-
-function getE2eCommitSearchResult(username: string): CommitData | null {
-  if (!E2E_COMMIT_SEARCH_MOCKS_ENABLED) return null;
-
-  // Keep browser tests deterministic without coupling them to GitHub search availability.
-  const mockCommit: CommitInfo = {
-    message: "Initial public commit\n\nAdd the first project files",
-    date: "2020-01-02T03:04:05Z",
-    html_url: "https://github.com/e2e-user/origin-repo/commit/abcdef123456",
-    sha: "abcdef123456",
-    repository: {
-      name: "origin-repo",
-      owner: "e2e-user",
-      full_name: "e2e-user/origin-repo",
-    },
-    author: {
-      login: username,
-      avatar_url: "https://github.com/ghost.png",
-      html_url: githubProfileUrl(username),
-    },
-  };
-
-  const completeResult = (): CommitData => ({
-    found: true,
-    commits: [
-      mockCommit,
-      {
-        ...mockCommit,
-        message: "Follow-up commit",
-        html_url: "https://github.com/e2e-user/next-repo/commit/bcdefa234567",
-        sha: "bcdefa234567",
-        repository: {
-          name: "next-repo",
-          owner: "e2e-user",
-          full_name: "e2e-user/next-repo",
-        },
-      },
-    ],
-  });
-
-  if (
-    username === "e2e-result" ||
-    username === "e2e-slow-result" ||
-    username.startsWith(E2E_REJECT_ONCE_USERNAME_PREFIX)
-  ) {
-    return completeResult();
-  }
-
-  // Every field at a length or shape that has nowhere to wrap: a repository name with no spaces,
-  // a subject that is one unbroken token, a body carrying a long URL, and a 39-character owner.
-  // The other fixtures are all short, so they render identically whether or not the card handles
-  // long content -- which is why the overflow this reproduces went unnoticed.
-  if (username === "e2e-long-data") {
-    const owner = "a".repeat(39);
-    const repositoryName = "supercalifragilisticexpialidocious-monorepo-services-platform";
-
-    return {
-      found: true,
-      commits: [
-        {
-          ...mockCommit,
-          message: [
-            "Refactorthecommitindexingpipelineandnormalizeeverydownstreamconsumer",
-            "",
-            "See https://github.com/" +
-              owner +
-              "/" +
-              repositoryName +
-              "/pull/12345#issuecomment-9876543210",
-          ].join("\n"),
-          repository: {
-            name: repositoryName,
-            owner,
-            full_name: owner + "/" + repositoryName,
-          },
-          author: { ...mockCommit.author, login: owner },
-        },
-        {
-          ...mockCommit,
-          message: "Follow-up with an unbroken subject line that keeps going and going and going",
-          html_url: "https://github.com/e2e-user/next-repo/commit/bcdefa234567",
-          sha: "bcdefa234567",
-          repository: {
-            name: repositoryName,
-            owner,
-            full_name: owner + "/" + repositoryName,
-          },
-          author: { ...mockCommit.author, login: owner },
-        },
-      ],
-    };
-  }
-
-  if (username === "e2e-incomplete") {
-    return {
-      found: true,
-      incomplete: true,
-      commits: [mockCommit],
-    };
-  }
-
-  // Stateful, so a browser test can prove a retry re-issues the search rather than
-  // re-rendering the same state: partial first, then the outcome the prefix names.
-  if (
-    username.startsWith(E2E_INCOMPLETE_ONCE_USERNAME_PREFIX) ||
-    username.startsWith(E2E_INCOMPLETE_THEN_ERROR_USERNAME_PREFIX)
-  ) {
-    if (!e2eServedPartialUsernames.has(username)) {
-      e2eServedPartialUsernames.add(username);
-      return {
-        found: true,
-        incomplete: true,
-        commits: [mockCommit],
-      };
-    }
-
-    if (username.startsWith(E2E_INCOMPLETE_THEN_ERROR_USERNAME_PREFIX)) {
-      return commitSearchError(
-        "GitHub rate limit reached. Please try again in a few minutes.",
-        "rate_limit",
-      );
-    }
-
-    return completeResult();
-  }
-
-  if (username === "e2e-malformed-dates") {
-    return {
-      found: true,
-      commits: [
-        mockCommit,
-        {
-          ...mockCommit,
-          message: "Commit with malformed date",
-          date: "not-a-date",
-          html_url: "https://github.com/e2e-user/next-repo/commit/bcdefa234567",
-          sha: "bcdefa234567",
-          repository: {
-            name: "next-repo",
-            owner: "e2e-user",
-            full_name: "e2e-user/next-repo",
-          },
-        },
-      ],
-    };
-  }
-
-  switch (username) {
-    case "e2e-empty":
-      return commitSearchError(EMPTY_COMMIT_SEARCH_MESSAGE, "empty");
-    case "e2e-incomplete-empty":
-      return commitSearchError(INCOMPLETE_COMMIT_SEARCH_MESSAGE, "empty", true);
-    case "e2e-rate-limit":
-      return commitSearchError(
-        "GitHub rate limit reached. Please try again in a few minutes.",
-        "rate_limit",
-      );
-    case "e2e-unavailable":
-      return commitSearchError(
-        "GitHub is temporarily unavailable. Please try again soon.",
-        "unavailable",
-      );
-    case "e2e-timeout":
-      return commitSearchError("GitHub took too long to respond. Please try again.", "timeout");
-    case "e2e-unknown":
-      return commitSearchError("GitHub commit search failed. Please try again.", "unknown");
-    // GitHub rejecting the search itself, which the client's own username validation cannot
-    // produce -- the handle is well-formed, and the refusal comes back from the search.
-    case "e2e-validation":
-      return commitSearchError("GitHub rejected that search.", "validation");
-    default:
-      return null;
-  }
 }
 
 function getGitHubErrorDetails(error: unknown): GitHubErrorDetails {
@@ -385,12 +201,7 @@ export async function getCommits(username: string): Promise<CommitData> {
   const validationMessage = getUsernameValidationMessage(normalizedUsername);
   if (validationMessage) return commitSearchError(validationMessage, "validation");
 
-  if (
-    E2E_COMMIT_SEARCH_MOCKS_ENABLED &&
-    normalizedUsername.startsWith(E2E_REJECT_ONCE_USERNAME_PREFIX) &&
-    !e2eRejectedUsernames.has(normalizedUsername)
-  ) {
-    e2eRejectedUsernames.add(normalizedUsername);
+  if (shouldRejectCommitSearchOnce(normalizedUsername)) {
     throw new Error("Synthetic E2E Server Action rejection");
   }
